requirements.txt
python-dotenv==1.2.2
rasterio==1.5.0
numpy==2.4.4
matplotlib==3.10.9